Added support for  PT2262 encoder
PT2262/CS2262 is a chip typically found in cheap no-brand hardware
such as alarm sensors and remote controls. The IC encodes a 12 bit tri-state
register of jumpers. Some times used as a 8 bit address and 4 bit code. In order
to keep decoding flexible, the implementation decodes the registers into a
single 10 base long value prior to transfer. Back-end decoding of that long
back to a ternary is needed in order to do processing of the register
values, or it could be used as a bell device with a unique identifier.